project_image_recognition/web/requirements.txt
Flask
flask_restful
pymongo
bcrypt
numpy
tensorflow
requests
Pillow
opencv-contrib-python-headless==3.4.3.18
